+++ demo/demoapp/cardlist/serializer.py
@@ -1,25 +1,10 @@
 from django.utils import formats
 from django.utils.translation import gettext_lazy as _
-from django_filters import DateFromToRangeFilter, DateTimeFromToRangeFilter
-from django_filters import FilterSet
 from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
 
 from demoapp.models import Person, Country
-from djgentelella.fields.drfdatetime import DateRangeTextWidget, DateTimeRangeTextWidget
 from djgentelella.serializers import GTDateField, GTDateTimeField
-
-
-class PersonFilterSet(FilterSet):
-    born_date = DateFromToRangeFilter(
-        widget=DateRangeTextWidget(attrs={'placeholder': 'YYYY/MM/DD'}))
-    last_time = DateTimeFromToRangeFilter(
-        widget=DateTimeRangeTextWidget(attrs={'placeholder': 'YYYY/MM/DD HH:MM:SS'}))
-
-    class Meta:
-        model = Person
-        fields = {'name': ['icontains'], 'num_children': ['exact'],
-                  'country__name': ['icontains']}
 
 
 class CountrySerializer(serializers.ModelSerializer):
@@ -47,6 +32,7 @@ class PersonCardListSerializer(serializers.Serializer):
     draw = serializers.IntegerField(required=True)
     recordsFiltered = serializers.IntegerField(required=True)
     recordsTotal = serializers.IntegerField(required=True)
+
 
 class PersonCreateSerializer(serializers.ModelSerializer):
     born_date = GTDateField()
