+++ demo/demoapp/templates/gentelella/cardlist/person_list.html
@@ -1,0 +1,18 @@
+{% extends 'gentelella/blocks/listcard_template.html' %}
+
+{% block listarea %}
+<% @each(it.data) => element %>
+<div class="card {{card_class}}">
+	<% @if(element.image != undefined) %>
+	<img src="<% element.image %>" class="card-img-top" alt="">
+	<% /if %>
+	<div class="card-body">
+		<h5 class="card-title"><% element.title %></h5>
+		<% element.body %>
+	</div>
+	{% if with_actions %}
+	<div class="card-footer action_area" id="{{id}}_action_area"></div>
+	{% endif %}
+</div>
+{{/each}}
+{% endblock %}
