+++ demo/demoapp/urls.py
@@ -6,11 +6,11 @@ from demoapp.cruds import Personclass, Countryclass, MenuItemclass
 from demoapp.views import create_notification_view, bt_modal_display
 from .autocomplete import views as autocompleteviews
 from .calendar.views import calendar_view
+from .cardlist.views import cardListViewExample
+from .cardlist.viewsets import PersonCardListViewSet
 from .chartjs import chart_js_view
 from .datatables.api import PersonViewSet
 from .datatables.views import datatableViewExample
-from .cardlist.views import cardListViewExample
-from .cardlist.api import PersonViewSet as PersonCard
 from .date_range import views as date_ranges
 from .filechunkedupload import views as chunkedupload
 from .formset import add_formset, add_model_formset
@@ -32,12 +32,11 @@ menuclss = MenuItemclass()
 
 router = DefaultRouter()
 router.register('persontableview', PersonViewSet, 'api-persontable')
-router.register('personcardview', PersonCard, 'api-personcard')
+router.register('personlistcardview', PersonCardListViewSet, 'api-personcard')
 router.register('objectmanagement', ObjectManagerDemoModelManagement,
                 'api-objectmanagement')
 
 urlpatterns = [
-
                   path('object_management', object_management,
                        name='object_management_index'),
                   path('bt_modal_display', bt_modal_display, name='bt_modal_display'),
@@ -110,5 +109,4 @@ urlpatterns = [
                   path('cardlist_view', cardListViewExample, name="cardlist_view"),
                   path('mediarecord_upload', mediaupload_view, name="mediaupload_view"),
                   path('tableapi/', include(router.urls)),
-                  path('cardapi/', include(router.urls)),
               ] + pclss.get_urls() + countryclss.get_urls() + menuclss.get_urls()
