+++ djgentelella/static/gentelella/js/base/cardList.js
