+++ djgentelella/templates/gentelella/blocks/squirrelly_pagination.html
@@ -1,0 +1,11 @@
+{% load i18n %}
+<nav aria-label="page navigation">
+	<ul class="pagination">
+		<li class="page-item"><a class="page-link" href="#" data-page="1">{% trans 'First' %}</a></li>
+		<% @each([...Array(it.totalPages).keys()]) => page %>
+		<li class="page-item <% page+1 == it.currentPage ? 'active' : '' %>">
+			<a class="page-link" href="#" data-page="<% page +1 %>"><% page+1 %></a></li>
+		<% /each %>
+		<li class="page-item"><a class="page-link" href="#" data-page="<% it.totalPages+1 %>">{% trans 'Last' %}</a></li>
+	</ul>
+</nav>
