--- djgentelella/templates/gentelella/widgets/list_template.html
@@ -1,76 +1,0 @@
-<!-- Evita que django reconozca el contenido permitiendole a squirrelly que haga uso de sus variables -->
-{% verbatim %}
-<div class="row container mt-3">
-    <!-- Filtros y búsqueda -->
-    <form id="filterForm" class="row" onsubmit="event.preventDefault(); cardList.fetchData();">
-        <div class="col-md-3">
-            <input type="text" id="searchInput" placeholder="Buscar..." class="form-control form-control-sm" onkeyup="cardList.filterData()">
-        </div>
-        <div class="col-md-3">
-            <input type="date" id="dateStart" class="form-control form-control-sm">
-        </div>
-        <div class="col-md-3">
-            <input type="date" id="dateEnd" class="form-control form-control-sm">
-        </div>
-        <div class="col-md-2">
-            <select id="sortField" class="form-select form-select-sm">
-                <option value="name">Nombre</option>
-                <option value="born_date">Fecha Nacimiento</option>
-            </select>
-        </div>
-        <div class="col-md-1">
-            <button type="submit" class="btn btn-sm btn-primary w-100">
-                <i class="fa fa-filter"></i>
-            </button>
-        </div>
-    </form>
-
-    <!-- Contenedor de Cards -->
-    <div class="row row-cols-1 row-cols-md-3 g-2 mb-2" id="cardList">
-
-        {{@each(it.data) => person}}
-					<div class="col-12 col-md-4 col-sm-12">
-        <div class="card">
-            <div class="card-body">
-                <h5 class="card-title">{{person.name}}</h5>
-                <p class="card-text">Hijos: {{person.num_children}}</p>
-                <p class="card-text">País: {{person.country.name}}</p>
-                <p class="card-text">Fecha de Nacimiento: {{person.born_date}}</p>
-            </div>
-        </div>
-						</div>
-        {{/each}}
-
-    </div>
-</div>
-
-<!-- Paginación -->
-<div class="row mt-2">
-    <div class="col-sm-12 col-md-5">
-        <div class="dataTables_info">
-            Mostrando {{(it.currentPage - 1) * 10 + 1}} a {{Math.min(it.currentPage * 10, it.recordsTotal)}} de {{it.recordsTotal}} registros
-        </div>
-    </div>
-    <div class="col-sm-12 col-md-7">
-        <ul class="pagination justify-content-end">
-            <li class="page-item {{it.currentPage == 1 ? 'disabled' : ''}}">
-                <a href="#" onclick="cardList.changePage(1)" class="page-link">Primero</a>
-            </li>
-            <li class="page-item {{it.currentPage == 1 ? 'disabled' : ''}}">
-                <a href="#" onclick="cardList.changePage({{it.currentPage - 1}})" class="page-link">Anterior</a>
-            </li>
-            {{@each([...Array(it.totalPages).keys()]) => page}}
-            <li class="page-item {{page+1 == it.currentPage ? 'active' : ''}}">
-                <a href="#" onclick="cardList.changePage({{page+1}})" class="page-link">{{page+1}}</a>
-            </li>
-            {{/each}}
-            <li class="page-item {{it.currentPage == it.totalPages ? 'disabled' : ''}}">
-                <a href="#" onclick="cardList.changePage({{it.currentPage + 1}})" class="page-link">Siguiente</a>
-            </li>
-            <li class="page-item {{it.currentPage == it.totalPages ? 'disabled' : ''}}">
-                <a href="#" onclick="cardList.changePage({{it.totalPages}})" class="page-link">Último</a>
-            </li>
-        </ul>
-    </div>
-</div>
-{% endverbatim %}
